move files to src directory
